requirements.txt
streamlit
PyPDF2
Pillow
pdf2image
streamlit-sortables
